Rename base request configuration file
