+++ kiota_abstractions/base_request_configuration.py
